pyproject.toml
[project]
name = "metroclaude"
version = "0.1.0"
description = "Pilot Claude Code from your phone via Telegram"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Ahmed", email = "ahmed@joyjuice.co" },
]
dependencies = [
    "python-telegram-bot>=21.0",
    "libtmux>=0.37.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
markdown = [
    "mistletoe",
    "telegramify-markdown",
]
voice = ["openai-whisper"]
screenshot = ["Pillow"]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
metroclaude = "metroclaude.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
